chore(all): prepare release 1.0.0-beta.2.0.1
This moves the shadow dom v1 slot feature detection into the mainline
channel. As a result, anyone using content element will now be using
emulated content elements since we detect the feature differently. This
should not be a breaking change to any code. It will just change the
runtime characteristics for those relying on the v0 behavior in Chrome.