fix(all): do not throw when globals are not defined
There is no way around aurelia-bootstrapper importing the pal-browser package. If we want to use the bootstrapper in the environment where globals like 'window' are not defined (e.g. NodeJS), we need to ensure this package does not throw any TypeErrors, even though we're not actually using any method from it in such environments.